requirements.txt
deepeval>=0.21.0
pandas>=1.5.0
google-generativeai>=0.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
pathlib2>=2.3.0
gradio>=4.0.0
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0